Update AGO-COMPONENT-RADIO-EXTNET.md
Enough bullshit for every millennia to come. I hope that this information helps people in Ukraine and Isreal and elsewhere unaware of it to disable and reconfigure their communications. More edits eventually. Fighting with a child and his mentally ill father to do the most basic things online and otherwise is impossible and pointless. Never again. Do not ever let these people live their lives free.